Split queue tree
